Variable Sinking & Climbing beep sounds, instead of fixed sounds
The sinking beep was programmed as a set value, which was played for every sinking rate above the sinking threshold. with this commit, the user can configure the sinking beep as it were before (fixed sound: change SINKING_BEEP_FREQ_UPDATE 0 in beeper.h) or set a sink sound update frequency in milliseconds. (currently set to 1 - change smoothly and rapidly)

Similarly, The Climbing beep was played at a fixed pitch. by changing the climbing beep frequency as it is playing, one gets a lot more information what the vario is doing. User can set this up the same way, with CLIMBING_BEEP_FREQ_UPDATE.